+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Given version `A.B.C.D`, breaking changes are to be expected in version number i
 - **http4k-connect-openfeature**: [New module] Standard OpenFeature Remote Evaluation Protocol client
 - **http4k-connect-openfeature-fake**: [New module] Fake to setup and evaluate OpenFeature flags
 - **http4k-ops-openfeature**: [New module] Support for plugging OpenFeature into request chain, including typesafe lenses for extracting flags from prinicpal and context during request processing.
+- **http4k-bridge-vertx**: Stream request and response bodies instead of buffering.
+- **http4k-bridge-ratpack**:  Stream request and response bodies instead of buffering.
 
 ### v6.48.0.0
 - **http4k-***: Upgrade versions